chore: initialize angular app with commit linting
Like the other repo with rotbox's API, the goal was to learn Angular.
The project shaped into something I had need for so I am splitting both
the back end and front end into their own repo. I had not bothered to
use git since it was more of a playground with no goal until now.