feat: setup basic html and ts for navbar
Setup the html to have all the navbar items and component logic to choose
what navbar items will be displayed if a user is logged in or not.

There is no logic currently to check if a user is logged in or not. The default
is that the user is not logged in.