feat: add router links and css to navbar
Add router links to the appropriate navigation links. Logout will not
navigate to anything on its own. It will perform actions so routing
must happen in a function. Add CSS to remove underlines on each link.
There should only be an underline on the active router link. Routing
wont be added here besides the names of the paths.